Move CLineupConfiguration traits
